pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "geoai-vlm"
version = "0.2.0"
description = "Geospatial Vision-Language Model analysis for street-level imagery. Download Mapillary images by location and generate structured descriptions using VLMs."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Yunus Serhat Bıçakçı", email = "yunus.serhat@marmara.edu.tr"},
    {name = "Yunus Serhat Bıçakçı", email = "yunusserhat.bicakci@glasgow.ac.uk"}
]
keywords = [
    "geoai",
    "street-view",
    "mapillary",
    "vlm",
    "vision-language-model",
    "urban-analytics",
    "geospatial",
    "geoparquet",
    "embedding",
    "clustering",
    "spatial-analysis",
    "vector-search",
    "multimodal",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"

dependencies = [
    # Core dependencies
    "geopandas>=1.1.2",
    "pandas>=2.3.3",
    "numpy<2.dev0",
    "shapely>=2.1.2",
    "pyarrow>=16.0.0",
    "haversine>=2.9.0",
    "tqdm>=4.67.0",
    "Pillow>=10.0.0",
    "requests>=2.32.5",
    "zensvi>=1.4.3",
    "vllm>=0.13.0",
    "qwen-vl-utils>=0.0.14",
    "transformers>=4.57.3",
    "torch>=2.9.0",
    "accelerate>=1.12.0",
    # Embedding & vector search
    "chromadb>=1.0.0",
    "faiss-cpu>=1.7.0",
    # Clustering & spatial analysis
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "libpysal>=4.9.0",
    "esda>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "black>=25.12.0",
    "ruff>=0.14.10",
    "mypy>=1.19.0",
]

[project.urls]
Homepage = "https://github.com/yunusserhat/geoai-vlm"
Documentation = "https://github.com/yunusserhat/geoai-vlm#readme"
Repository = "https://github.com/yunusserhat/geoai-vlm"
Issues = "https://github.com/yunusserhat/geoai-vlm/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/geoai_vlm"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D", "UP"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
[tool.uv]
override-dependencies = [
    "huggingface-hub<1.0.0",
    "attrs>=22.2.0",
]